requirements.txt
astroid==3.3.9
certifi==2025.4.26
charset-normalizer==3.4.2
chemparse==0.3.2
colorama==0.4.6
contourpy==1.3.1
coverage==7.7.1
cycler==0.12.1
dill==0.3.9
et_xmlfile==2.0.0
exceptiongroup==1.2.2
fonttools==4.56.0
idna==3.10
iniconfig==2.1.0
isort==6.0.1
kiwisolver==1.4.8
libcellml==0.6.3
libChEBIpy==1.0.10
lxml==6.0.0
matplotlib==3.10.1
mccabe==0.7.0
mpmath==1.3.0
nose==1.3.7
numpy==2.2.6
openpyxl==3.1.5
packaging==24.2
pandas==2.3.1
pillow==11.1.0
platformdirs==4.3.7
pluggy==1.5.0
pylint==3.3.6
pyparsing==3.2.3
pytest==8.3.5
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-libsbml==5.20.5
pytz==2025.2
requests==2.32.3
SBMLKinetics==1.1.1
scipy==1.15.2
seaborn==0.13.2
six==1.17.0
sympy==1.14.0
tomli==2.2.1
tomlkit==0.13.2
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
XlsxWriter==3.2.2
